붙을 수 있는 오브젝트가 플레이어에게 붙는 동작을 할 때, 플레이어에게 끌려옴(플레이어 날아감)

+++ Assets/Resources/Scripts/Player/GlueableObj.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class GlueableObj : MonoBehaviour
 {
     private Collider myCol = null;
-    public double size { get; private set; }
+    public float size { get; private set; }
 
     void Start()
     {
